+++ srcs/push_swap/ft_double_insert_sort.c
@@ -6,7 +6,7 @@
 /*   By: gcros <gcros@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/10 03:42:45 by gcros             #+#    #+#             */
-/*   Updated: 2025/02/09 18:40:24 by gcros            ###   ########.fr       */
+/*   Updated: 2025/02/09 23:08:40 by gcros            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,9 +37,9 @@ int	ft_double_insert_sort_brain(t_sort *sort)
 	while (sort->stack_a->size > 0)
 	{
 		n = ft_find_value(sort->stack_a,
-				((long *)sort->stack_b->data)[sort->stack_b->size - 1] - 1);
+				((int *)sort->stack_b->data)[sort->stack_b->size - 1] - 1);
 		tmp_n = ft_find_value(sort->stack_a,
-				((long *)sort->stack_b->data)[0] + 1);
+				((int *)sort->stack_b->data)[0] + 1);
 		if (ft_get_dist(sort->stack_a, n) > ft_get_dist(sort->stack_a, tmp_n))
 			n = tmp_n;
 		if (ft_goto_a(sort, n) == -1)
@@ -47,7 +47,7 @@ int	ft_double_insert_sort_brain(t_sort *sort)
 		if (ft_run_op(sort, PB))
 			if (ft_add_operator(sort->instruction, PB) == -1)
 				return (-1);
-		if (((long *)sort->stack_b->data)[0] < ((long *)sort->stack_b->data)[1])
+		if (((int *)sort->stack_b->data)[0] < ((int *)sort->stack_b->data)[1])
 			if (ft_run_op(sort, RB))
 				if (ft_add_operator(sort->instruction, RB) == -1)
 					return (-1);
